+++ tests/Database/Factories/MaterialFactory.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Spiritix\LadaCache\Tests\Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
@@ -10,6 +12,9 @@ class MaterialFactory extends Factory
     /** @var class-string<\Spiritix\LadaCache\Tests\Database\Models\Material> */
     protected $model = Material::class;
 
+    /**
+     * @inheritDoc
+     */
     public function definition(): array
     {
         return [