+++ tests/Database/Models/Car.php
@@ -1,9 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Spiritix\LadaCache\Tests\Database\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasOne;
 use Spiritix\LadaCache\Database\LadaCacheTrait;
 
 class Car extends Model
@@ -12,17 +17,17 @@ class Car extends Model
 
     protected $fillable = ['name', 'engine_id', 'driver_id'];
 
-    public function engine(): \Illuminate\Database\Eloquent\Relations\HasOne
+    public function engine(): HasOne
     {
         return $this->hasOne(Engine::class);
     }
 
-    public function driver(): \Illuminate\Database\Eloquent\Relations\BelongsTo
+    public function driver(): BelongsTo
     {
         return $this->belongsTo(Driver::class);
     }
 
-    public function materials(): \Illuminate\Database\Eloquent\Relations\BelongsToMany
+    public function materials(): BelongsToMany
     {
         return $this->belongsToMany(Material::class);
     }