+++ tests/Database/Models/Driver.php
@@ -1,9 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Spiritix\LadaCache\Tests\Database\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Spiritix\LadaCache\Database\LadaCacheTrait;
 
 class Driver extends Model
@@ -12,7 +15,7 @@ class Driver extends Model
 
     protected $fillable = ['name'];
 
-    public function cars(): \Illuminate\Database\Eloquent\Relations\HasMany
+    public function cars(): HasMany
     {
         return $this->hasMany(Car::class);
     }