+++ tests/Database/Models/Engine.php
@@ -1,9 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Spiritix\LadaCache\Tests\Database\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Spiritix\LadaCache\Database\LadaCacheTrait;
 
 class Engine extends Model
@@ -12,7 +15,7 @@ class Engine extends Model
 
     protected $fillable = ['name', 'car_id'];
 
-    public function car(): \Illuminate\Database\Eloquent\Relations\BelongsTo
+    public function car(): BelongsTo
     {
         return $this->belongsTo(Car::class);
     }