+++ tests/Database/Models/Material.php
@@ -1,9 +1,11 @@
 <?php
+declare(strict_types=1);
 
 namespace Spiritix\LadaCache\Tests\Database\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Spiritix\LadaCache\Database\LadaCacheTrait;
 
 class Material extends Model
@@ -12,7 +14,7 @@ class Material extends Model
 
     protected $fillable = ['name'];
 
-    public function cars(): \Illuminate\Database\Eloquent\Relations\BelongsToMany
+    public function cars(): BelongsToMany
     {
         return $this->belongsToMany(Car::class);
     }