+++ tests/TestCase.php
@@ -1,52 +1,50 @@
 <?php
+declare(strict_types=1);
 
 namespace Spiritix\LadaCache\Tests;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Facades\DB;
+use Orchestra\Testbench\TestCase as BaseTestCase;
 use Spiritix\LadaCache\LadaCacheServiceProvider;
 
-class TestCase extends \Orchestra\Testbench\TestCase
+class TestCase extends BaseTestCase
 {
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
 
         $migrationParams = [
             '--database' => 'testing',
-            '--path' => realpath(__DIR__.'/../database/migrations'),
+            '--path' => realpath(__DIR__.'/Database/migrations'),
             '--realpath' => true,
         ];
 
         $this->artisan('migrate', $migrationParams);
 
-        Factory::guessFactoryNamesUsing(static function (string $modelName) {
-            return sprintf("Spiritix\\LadaCache\\Tests\\Database\\Factories\\%sFactory", class_basename($modelName));
+        Factory::guessFactoryNamesUsing(static function (string $modelName): string {
+            return sprintf('Spiritix\\LadaCache\\Tests\\Database\\Factories\\%sFactory', class_basename($modelName));
         });
 
         DB::beginTransaction();
     }
 
-    public function tearDown(): void
+    protected function tearDown(): void
     {
-        DB::rollback();
+        DB::rollBack();
 
         parent::tearDown();
     }
 
-    protected function getPackageProviders($app)
+    protected function getPackageProviders($app): array
     {
         return [
             LadaCacheServiceProvider::class
         ];
     }
 
-    protected function getEnvironmentSetUp($app)
+    protected function getEnvironmentSetUp($app): void
     {
-        $app['config']->set('lada-cache.active', true);
-        $app['config']->set('lada-cache.prefix', 'lada:');
-        $app['config']->set('lada-cache.consider_rows', true);
-
         $app['config']->set('database.default', 'testing');
         $app['config']->set('database.connections.testing', [
             'driver'   => 'sqlite',